pyproject.toml
[project]
name = "TimescaleDB"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Nikita", email = "savvin.nikit@yandex.ru"},
]
dependencies = [
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "TimescaleDB @ file:///${PROJECT_ROOT}/",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
